Just a quick tear down of the Transforms.js input component.
In preparation for removing canvas calls from the turtle.js port, and appending vertices to the MachinePreview.

Suggestion: BaseInput.js as a starting point for new input components.

I have produced a rough and ready one in this commit.  If it's a agreed a base input component would be useful, I would welcome some guidance or discussion about what characteristics this would have.